LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/getallresponseheaders-cl-expected.txt


PASS Casing of known headers 
FAIL Casing of known headers 1 assert_regexp_match: expected object "/content-TYPE/" but got ""
FAIL Casing of known headers 2 assert_regexp_match: expected object "/THIS-is-A-test: 1, 2/" but got ""
